Split separate mailing address input
