+++ packages/client/src/comp/Address/Form.tsx
@@ -1,61 +1,24 @@
 import React from 'react'
-import Input from 'muicss/lib/react/input'
-
-import { RoundedButton } from '../util/Button'
 import { client } from '../../lib/trpc'
 import { AddressContainer, ContactContainer, FetchingDataContainer } from '../../lib/unstated'
-import { useControlRef } from '../util/ControlRef'
 import { TimeoutError } from '@tianhuil/simple-trpc/dist/timedFetch'
-import { BaseInput } from '../util/Input'
 import { StatusReport } from '../status/StatusReport'
 import { useParams } from 'react-router-dom'
 import { useAppHistory } from '../../lib/path'
-import styled from 'styled-components'
-import { getState, allStates } from '../../common'
+import { getState } from '../../common'
 import { AppForm } from '../util/Form'
 import { Unidentified } from '../status/Status'
 import { toast } from 'react-toastify'
 import { AddressInputPartContainer } from '.'
-import { Select, Option } from 'muicss/react'
-
-const FlexBox = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  flex-flow: row wrap;
-  align-items: center;
-  margin: 1em 0;
-`
-
-const Flex = styled.div<{ basis?: string }>`
-  flex-basis: ${p => p.basis ?? '100%'};
-  /* Fixes needless scrollbars showing on Linux/Windows */
-  .mui-select label { overflow: hidden; }
-`
-
-// Since react-hooks/exhaustive-deps will trigger re-renders due to its
-// depedencies, we declare this function outside of our component.
-const useDidMount = (fun: () => void) => React.useEffect(fun, [])
+import { AddressFields } from './Fields'
 
 // pulled out for testing
 export const RawAddressForm: React.FC<{rawState: string, zip?: string}> = ({rawState}) => {
-  const { fields, setField } = AddressInputPartContainer.useContainer()
-  const { path, pushState } = useAppHistory()
-  const addrRef = useControlRef<Input>()
+  const { fields } = AddressInputPartContainer.useContainer()
+  const { pushState } = useAppHistory()
   const { setAddress } = AddressContainer.useContainer()
   const { setContact } = ContactContainer.useContainer()
-  const { fetchingData, setFetchingData } = FetchingDataContainer.useContainer()
-
-
-  // When we first arrive at page, set focus and move cursor to beginning
-  useDidMount(() => {
-    if (path?.type === 'address' && addrRef?.current) {
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      const controlEl = (addrRef.current as any).controlEl as HTMLInputElement
-      controlEl.focus({preventScroll: true})
-      controlEl.setSelectionRange(0, 0)
-    }
-  })
+  const { setFetchingData } = FetchingDataContainer.useContainer()
 
   const state = getState(rawState)
   if (!state) {
@@ -66,8 +29,6 @@ export const RawAddressForm: React.FC<{rawState: string, zip?: string}> = ({rawS
     event.persist()  // allow async function call
     event.preventDefault()
 
-    const addr = addrRef.value()
-    if (addr === null) throw Error('address ref not set')
     if (!state) throw Error('This can never happen: already checked if state is valid')
 
     setFetchingData(true)
@@ -109,100 +70,7 @@ export const RawAddressForm: React.FC<{rawState: string, zip?: string}> = ({rawS
         Enter your <b>voter-registration address</b> to find your local election official.
         (<a target='_blank' rel='noopener noreferrer' href='https://www.vote.org/am-i-registered-to-vote/'>Unsure if you are registered?</a>)
       </p>
-      <FlexBox>
-        {/* We use a 7:3 basis proportion between Flex Components */}
-        {/* The spacing between them 2% of the basis */}
-
-        <Flex basis='70%'>
-          <BaseInput
-            id='addr-street-input'  // This id is used for Warning Box to fill form quickly
-            label='Street Address'
-            ref={addrRef}
-            required
-            translate='no'
-            lang='en'
-            value={fields.street}
-            onChange={e => setField('street', e.currentTarget.value)}
-          />
-        </Flex>
-
-        <Flex basis='28%'>
-          <BaseInput
-            label='Apartment'
-            id='addr-apt-input'
-            translate='no'
-            lang='en'
-            value={fields.unit}
-            onChange={e => setField('unit', e.currentTarget.value)}
-          />
-        </Flex>
-
-        <Flex basis='40%'>
-          <BaseInput
-            id='addr-city-input'  // This id is used for Warning Box to fill form quickly
-            label='City'
-            required
-            translate='no'
-            lang='en'
-            value={fields.city}
-            onChange={e => setField('city', e.currentTarget.value)}
-          />
-        </Flex>
-
-        <Flex basis='28%' className='mui-select'>
-          <Select
-            id='addr-state-input'  // This id is used for Warning Box to fill form quickly
-            label='State'
-            translate='no'
-            lang='en'
-            onChange={e => {
-              // MuiCSS has a buggy support for <Select/> when using TypeScript,
-              // to really access the HTMLSelect and its value we need
-              // to do this hack
-              const trueSelect = e.currentTarget.firstChild as HTMLSelectElement
-              setField('state', trueSelect.value)
-            }}
-          >
-            {
-              [...allStates].sort().map((state, key) => {
-                return <Option
-                  key={`${state}${key}`}
-                  value={state}
-                  selected={state === fields.state}
-                  label={state}
-                >
-                  {state}
-                </Option>
-              })
-            }
-          </Select>
-        </Flex>
-
-        <Flex basis='28%'>
-          <BaseInput
-            id='addr-zip-input'  // This id is used for Warning Box to fill form quickly
-            label='ZIP code'
-            required
-            translate='no'
-            lang='en'
-            value={fields.postcode}
-            onChange={e => setField('postcode', e.currentTarget.value)}
-          />
-        </Flex>
-
-        <Flex>
-          <RoundedButton
-            id='addr-submit'  // This id is used for Warning Box to submit form quickly
-            color='primary'
-            variant='raised'
-            data-testid='submit'
-            style={{flexGrow: 0}}
-            disabled={fetchingData}
-          >
-            Find my election official
-          </RoundedButton>
-        </Flex>
-      </FlexBox>
+      <AddressFields type='initialAddress'/>
     </AppForm>
   </StatusReport>
 }
