+++ packages/client/src/comp/Address/index.tsx
@@ -1,2 +1,3 @@
 export * from './Container'
 export * from './Form'
+export * from './Fields'
